Retry block in `future_stencil` to prevent loading incomplete modules (#616)
This PR implements a fix in `future_stencil` to prevent the loading of incomplete modules. This resolves intermittent fails when running the `numpy` backend with 54 ranks. The lack of compilation means `numpy` modules are generated quickly and can potentially "exist" before the other node has finished writing the file. During that time another node may attempt to load the file resulting in an attribute error.

Code changes:
* Add retry try/catch block to module load in `future_stencil._wait_for_stencil`