Add dependencies graph for project to README (#8)
We plan to modify the testing framework so that tests for a project run when any of its dependencies change. Doing this requires it to be very clear what should and should not depend on what else. This PR adds documentation for the interdependencies of the packages in this repo.