Bugfix/c48 on daint (#634)
A few more pieces in place to be able to run c48_6ranks standard on daint in jenkins. The reporting of the max error could be incorrect in the case of ignore_near_zero -- it was reporting points that weren't actually flagged as failures because the values were so tiny. This is less fancy, but uses the identified errors. Updated thresholds for D_SW and MapN_Tracer_2d that seem necessary for c48_6ranks_standard tests.