Remove global stencil config from FrozenStencil using StencilFactory (#632)
## Purpose

Currently, FrozenStencil retrieves a global StencilConfig object when initializing a new stencil. This PR removes the global default, which means StencilConfig must always be passed explicitly.

To pass StencilConfig between our classes, we add a new StencilFactory class which contains StencilConfig and GridIndexing, and use it to create our stencils. It passes its stencil configuration to FrozenStencil in its factory methods.

The StencilFactory class also includes a method to initialize a stencil directly from dimensions and halo counts, condensing a pattern that appears throughout the code base. In this PR, several files are refactored to use this pattern.

## Code changes:

- Created StencilFactory class
- FrozenStencil now has stencil_config as a required argument
- Initialization methods throughout the repo now pass a StencilFactory instead of a GridIndexing class, and retrieve the GridIndexing class from this factory.
- The following modules in fv3core.stencils have been refactored to make some use of from_dims_halos: c_sw, d_sw, c2l_ord4, a2b_ord4, d2a2c_vect, del2cubed, and divergence_damping