Make jenkins green... again (#35)
Fixing various issues that are disrupting our tests, including:

replace cached_property in MetricTerms with a decorator that works, as the former was caching at the class level, not the instance level, impacting validation when the tests are run together
adjusting some thresholds that need changing for running standalone, for grid computation on c48
add a --skip_grid_tests flag to test_regressions that enables turning off testing grid generation. This was added so on PRs we can not run these tests (since the code is unlikely to change, and if impacted by other changes, other code is to likely fail as well) because they increase the time to compute (especially on c54 ranks) significantly. We will not turn off testing the grid in the regressions-cron plan so that someone is alerted if a code change breaks grid generation
fixed the state initialization to work on gpu, doing numpy operations on the memoryview does not work
run_on_daint script takes a namelist and an experiment name rather than a whole dataset, since we don't need the serialized data for this script anymore. run_standalone passes the namelist and experiment name to it.