init baroclinic gpu (#67)
On GPU the initilized state was actually all 0s, as the effort to keep the code from crashing on GPU actually entirely separated the gpu arrays from the cpu arrays. It is possible to pass cupy arrays (the state.quantity.data instead of the state.storage.data, which is a memory view and cannot be sliced) in and still do it this way, but this requires modifying almost all the lines of math to be cupy friendly (e.g. cp.asarray wrappers around many statements). Instead, we compute on the cpu in numpy, and from that create a DycoreState.
This means we won't use the quantity.view feature to slice unless we revisit using a a DycoreState initializaed with 0's to start with