add pace driver, update dycore and physics state (#79)
Purpose
Added a 'pace-driver' package for use in managing model workflows.

Code changes
Added a simple Driver class that is meant as a stepping-stone, not the final design, to move the complexity of objects being created in runfiles into the 'model.'
The driver can use serialized initial state data or baroclinic
Updated the PhysicsState to use similar conventions as the DycoreState, and include metadata that indicates whether it is shared with dycore (full_model_var). In this version shared variables are not copied, but shared. We may want to adjust and pull up stencil code that transforms dycore variables to physics variables and vice versa up to the Driver level, but that is likely better as a future iteration.
MicrophysicsState doesn't initialize storages in call-time
Moved fv_update_phys, update_atmos_state, and update_dwind_phys into pace-stencils, and the state updater object into the Driver.
u_dt, v_dt, and t_dt in the updater are quantities
remove 'partitioner' as a separate argument from the communicator in Physics classes because we can get the partitioner from the communicator.
created a DriverGridData class that extracts the variables update_dwind_phys needs from MericTerms. We no longer need serialized grid data for pace (except for testing regressions purposes).
Physics no longer refers to DycoreState
removed hard-coded ptop from Physics -- it is available in grid_data
added examples/run_driver.py as a script that demonstrates the Driver class. We could alternatively adapt examples/dyn_phy.py and remove this file.
Namelist gets the dycore_only configuration option turned on so we can use it.
MetricTerms has a new method for splitting cartesian variables into FloatIJ storages, as seems like it could be a common use-case for someone using MetricTerms.
Updated test code to go with the changes