update making the DycoreState from serialized data and add a test endpoint in the driver Makefile so it is easy to ensure it still works. (#96)
Purpose
With the Driver we added the option to initialize from serialized data, because there are experiments we want to run that don't start from the baroclinic analytic state init. We didn't add a test for this functionality, however, and a recent change broke it. This fixes it and adds a driver test that just checks run.py can run, does not do regressions yet (coming soon)

Code changes:
update method in the fv3core/testing/translate_fvdynamics code that makes a DycoreState from serialized data
update MAX_SIZE for future stencil to accommodate dycore+physics+driver stencils
synchronization of the state before writing out is fixed to synchronize the storages now that the state fields are storages
Infrastructure changes:
added test_serialized_init to the driver Makefile that checks the driver/run.py script does not crash when starting from serialized data