single node compilation (#133)
Purpose
Since the daint upgrade, we haven't been able to compile multi-node, which could either be due to an issue with our distributed compilation, or a bug in the new cuda stack (gcc 10 no longer works with cuda 11.2).
This introduces a dycore compiling script compile.py that will take the namelist we are testing and compile all the ranks in the first tile sequentially. For 6 ranks we can overload a node and do single node compilation, but this turned out to be slower than running them sequentially. But if we want to avoid running compile.py in our CI, I can branch on the number of ranks.

Code changes:
added compile.py that runs the dycore with a null communicator and remapping replaced (to avoid the while loop running off the top of data that is nans)
updated run_on_daint to first run compile.py, then run the performance test.