Remove SerialboxConfig initialization configuration from pace.driver (#168)
## Purpose

Previously we had created a SerialboxConfig initialization configuration to validate the output of the driver on Fortran regression cases. In the end we opted for a Translate test to check values which does not make use of this configuration, and instead uses the predefined-state configuration. This PR removes the unused configuration and its test.