Zero heat_source temporary between dycore calls (#196)
## Purpose

Due to heat_source not being re-initialized to zero at the appropriate point where it is done in the Fortran code, the dycore was retaining model state and producing incorrect behavior on subsequent timesteps. This PR fixes that issue. It also adds a unit test which ensures temporaries kept as attributes on our objects do not retain model state between calls. It does this by ensuring that if a dycore is called twice with the same model state, the resulting temporaries should be identical.

## Code changes:

- Fixed bug where heat_source and diss_est were not zeroed when first used in D_SW
- Added test that if a dycore is called twice with the same model state, the resulting temporaries should be identical.
- lagrangian contributions is no longer excluded from dycore main tests, since it no longer causes an exception on the numpy backend
- init is no longer an argument to `make_storage_from_shape`, instead it always initializes zeros