Use netCDF data loading in physics tests (#209)
## Purpose

This PR refactors the physics and driver translate tests to use the logic from fv3core for netCDF data loading.

This PR introduces some fairly unclear import logic which requires a lot of flake8 ignore statements. In a later PR, we should refactor the tests to all lie under the same directory and share the same testing code. This was left out in this PR to limit its scope.