restore dynamical core statefulness test (#234)
## Purpose

A previous PR disabled the test that our dynamical core is not stateful due to needing to bind the dynamical core to a particular state for halo updates. This PR restores that test by altering it to copy data into that single bound state instance, instead of calling the dynamical core with two separate state instances.